load files by chunks, this allows us to open big files, no matter how big they are. plus automatically reload file when cursor is at edge of current loaded chunk

+++ src/main/java/roksard/typingTrainer/pojo/Config.java
@@ -19,7 +19,7 @@ public class Config {
     private Integer winW;
     private Integer winH;
     private String fileName;
-    private Integer filePos;
+    private Long filePos;
     private List<Statistic> statistic = new ArrayList<>();
     String fontName;
     private Integer fontStyle;
